+++ CHANGELOG.md
@@ -54,6 +54,9 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 ### Changed
 
+- SqlDatabaseUser
+  - The parameter `ServerName` is now non-mandatory and defaults to `$env:COMPUTERNAME`
+    ([issue #319](https://github.com/dsccommunity/SqlServerDsc/issues/319)).
 - SqlServerConfiguration
   - The parameter `ServerName` is now non-mandatory and defaults to `$env:COMPUTERNAME`
     ([issue #319](https://github.com/dsccommunity/SqlServerDsc/issues/319)).